requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.43.3
Babel==2.14.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
defusedxml==0.7.1
Eel==0.16.0
et-xmlfile==1.1.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.6
lxml==5.1.1
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pefile==2023.2.7
pillow==10.2.0
pycparser==2.22
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.31.0
six==1.16.0
tkcalendar==1.6.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.2
